move TASK_COMPILE_SOLIDITY
