+++ packages/hardhat-core/src/build-system/build-system.ts
@@ -50,6 +50,7 @@ import chalk from "chalk";
 // import { DependencyGraph } from "../internal/solidity/dependencyGraph";
 
 // ------------------ START PACKAGE
+import { CompilationJobsCreationResult } from "../types/builtin-tasks";
 import { DependencyGraph } from "./util/solidity/dependencyGraph";
 import {
   Artifacts,
@@ -88,6 +89,8 @@ import {
 import { getEvmVersionFromSolcVersion } from "./util/solidity/compiler/solc-info";
 import { Parser } from "./util/solidity/parse";
 import { Resolver } from "./util/solidity/resolver";
+import { getSolidityFilesCachePath } from "./util/solidity-files-cache";
+import { getFullyQualifiedName } from "./util/contract-names";
 // --------------END PACKAGE
 
 const log = debug("hardhat:core:tasks:compile:REFACTORING");
@@ -1260,7 +1263,166 @@ export async function taskCompileSolidityGetDependencyGraph(
  */
 export async function taskCompileTransformImportName(
   importName: string,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   deprecationCheck: boolean
 ): Promise<string> {
   return importName;
+}
+
+// TASK_COMPILE_SOLIDITY
+// TESTED
+/**
+ * Main task for compiling the solidity files in the project.
+ *
+ * The main responsibility of this task is to orchestrate and connect most of
+ * the subtasks related to compiling solidity.
+ */
+export async function taskCompileSolidity(
+  config: HardhatConfig,
+  artifacts: Artifacts,
+  force: boolean,
+  quiet: boolean,
+  concurrency: number,
+  tasksOverrides: TasksOverrides
+) {
+  const rootPath = config.paths.root;
+
+  const sourcePaths: string[] = await taskCompileSolidityGetSourcePaths(
+    config,
+    config.paths.sources
+  );
+
+  const sourceNames: string[] = await taskCompileSolidityGetSourceNames(
+    config,
+    sourcePaths,
+    rootPath
+  );
+
+  const solidityFilesCachePath = getSolidityFilesCachePath(config.paths);
+  let solidityFilesCache = await SolidityFilesCache.readFromFile(
+    solidityFilesCachePath
+  );
+
+  const dependencyGraph: taskTypes.DependencyGraph =
+    await taskCompileSolidityGetDependencyGraph(
+      sourceNames,
+      config,
+      tasksOverrides,
+      rootPath,
+      solidityFilesCache
+    );
+
+  solidityFilesCache = await invalidateCacheMissingArtifacts(
+    solidityFilesCache,
+    artifacts,
+    dependencyGraph.getResolvedFiles()
+  );
+
+  const compilationJobsCreationResult: CompilationJobsCreationResult =
+    await taskCompileSolidityGetCompilationJobs(
+      config,
+      dependencyGraph,
+      solidityFilesCache
+    );
+
+  await taskCompileSolidityHandleCompilationJobsFailures(
+    compilationJobsCreationResult.errors
+  );
+
+  const compilationJobs = compilationJobsCreationResult.jobs;
+
+  const filteredCompilationJobs: CompilationJob[] =
+    await taskCompileSolidityFilterCompilationJobs(
+      compilationJobs,
+      force,
+      solidityFilesCache
+    );
+
+  const mergedCompilationJobs: CompilationJob[] =
+    await taskCompileSolidityMergeCompilationJobs(filteredCompilationJobs);
+
+  const {
+    artifactsEmittedPerJob,
+  }: { artifactsEmittedPerJob: ArtifactsEmittedPerJob } =
+    await taskCompileSolidityCompileJobs(
+      mergedCompilationJobs,
+      artifacts,
+      quiet,
+      concurrency
+    );
+
+  // update cache using the information about the emitted artifacts
+  for (const {
+    compilationJob: compilationJob,
+    artifactsEmittedPerFile: artifactsEmittedPerFile,
+  } of artifactsEmittedPerJob) {
+    for (const { file, artifactsEmitted } of artifactsEmittedPerFile) {
+      solidityFilesCache.addFile(file.absolutePath, {
+        lastModificationDate: file.lastModificationDate.valueOf(),
+        contentHash: file.contentHash,
+        sourceName: file.sourceName,
+        solcConfig: compilationJob.getSolcConfig(),
+        imports: file.content.imports,
+        versionPragmas: file.content.versionPragmas,
+        artifacts: artifactsEmitted,
+      });
+    }
+  }
+
+  const allArtifactsEmittedPerFile = solidityFilesCache.getEntries();
+
+  // We know this is the actual implementation, so we use some
+  // non-public methods here.
+  const artifactsImpl = artifacts as ArtifactsImpl;
+  artifactsImpl.addValidArtifacts(allArtifactsEmittedPerFile);
+
+  await solidityFilesCache.writeToFile(solidityFilesCachePath);
+
+  if (tasksOverrides?.taskCompileSolidityLogCompilationResult !== undefined) {
+    await tasksOverrides.taskCompileSolidityLogCompilationResult(
+      mergedCompilationJobs,
+      quiet
+    );
+  } else {
+    await taskCompileSolidityLogCompilationResult(mergedCompilationJobs, quiet);
+  }
+}
+
+/**
+ * If a file is present in the cache, but some of its artifacts are missing on
+ * disk, we remove it from the cache to force it to be recompiled.
+ */
+async function invalidateCacheMissingArtifacts(
+  solidityFilesCache: SolidityFilesCache,
+  artifacts: Artifacts,
+  resolvedFiles: ResolvedFile[]
+): Promise<SolidityFilesCache> {
+  const paths = new Set(await artifacts.getArtifactPaths());
+
+  for (const file of resolvedFiles) {
+    const cacheEntry = solidityFilesCache.getEntry(file.absolutePath);
+
+    if (cacheEntry === undefined) {
+      continue;
+    }
+
+    const { artifacts: emittedArtifacts } = cacheEntry;
+    for (const emittedArtifact of emittedArtifacts) {
+      const fqn = getFullyQualifiedName(file.sourceName, emittedArtifact);
+      const path = artifacts.formArtifactPathFromFullyQualifiedName(fqn);
+
+      if (!paths.has(path)) {
+        log(
+          `Invalidate cache for '${file.absolutePath}' because artifact '${fqn}' doesn't exist`
+        );
+
+        solidityFilesCache.removeEntry(file.absolutePath);
+        break;
+      }
+    }
+  }
+
+  artifacts.clearCache?.();
+
+  return solidityFilesCache;
 }
